100-get_endianness.c

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -7,7 +7,7 @@
 
 int get_endianness(void)
 {
-	unsigned int integer = 1;
+	unsigned long int integer = 1;
 	char *character = (char *)&integer;
 
 	return (*character);
